+++ bw.core/src/main/java/sa/gov/nic/bio/bw/core/utils/AppConstants.java
@@ -26,8 +26,9 @@ public interface AppConstants
 	String UI_LANGUAGE_PREF_NAME = "sa.gov.nic.bio.bw.ui.language";
 	String LOGIN_FINGERPRINT_POSITION_PREF_NAME = "sa.gov.nic.bio.bw.login.fingerprint";
 	ZoneId SAUDI_ZONE = ZoneId.of("GMT+3");
-	String LOGS_FOLDER_PATH = "C:/bio/user-apps/" + System.getProperty("user.name") + "/bw/logs";
-	String TEMP_FOLDER_PATH = "C:/bio/user-apps/" + System.getProperty("user.name") + "/bw/temp";
+	String APP_FOLDER_PATH = "C:/bio/user-apps/" + System.getProperty("user.name") + "/bw";
+	String LOGS_FOLDER_PATH = APP_FOLDER_PATH + "/logs";
+	String TEMP_FOLDER_PATH = APP_FOLDER_PATH + "/temp";
 	String DEV_SERVER_URL = "10.0.73.80:8080"; // TODO: TEMP
 	KeyCombination SCENIC_VIEW_KEY_COMBINATION = new KeyCodeCombination(KeyCode.D,
 	                                                                    KeyCombination.CONTROL_DOWN,
@@ -38,6 +39,10 @@ public interface AppConstants
                                                                                KeyCombination.SHIFT_DOWN,
                                                                                KeyCombination.ALT_DOWN);
 	KeyCombination CHANGING_SERVER_KEY_COMBINATION = new KeyCodeCombination(KeyCode.S,
+	                                                                        KeyCombination.CONTROL_DOWN,
+	                                                                        KeyCombination.SHIFT_DOWN,
+	                                                                        KeyCombination.ALT_DOWN);
+	KeyCombination OPEN_APP_FOLDER_KEY_COMBINATION = new KeyCodeCombination(KeyCode.A,
 	                                                                        KeyCombination.CONTROL_DOWN,
 	                                                                        KeyCombination.SHIFT_DOWN,
 	                                                                        KeyCombination.ALT_DOWN);