+++ bw.workflow.deletecompletecriminalrecord/src/main/java/sa/gov/nic/bio/bw/workflow/deletecompletecriminalrecord/utils/DeleteCompleteCriminalRecordErrorCodes.java
@@ -2,7 +2,9 @@ package sa.gov.nic.bio.bw.workflow.deletecompletecriminalrecord.utils;
 
 public enum DeleteCompleteCriminalRecordErrorCodes
 {
-	C015_00001;
+	C015_00001,
+	
+	N015_00001;
 	
 	public final String getCode()
 	{